moved the svgs into their own files
